requirements.txt
PyGithub>=2.0.0
requests>=2.31.0
